+++ todoapp/app/javascript/packs/application.js
@@ -9,29 +9,11 @@
 
 import '../../src/application.scss'
 import 'bootstrap/dist/js/bootstrap'
+import Confirm from './components/confirm'
 
 /*
  * RailsでjQueryを読み込まずにmethod: :deleteのリンクを実現する（confirmの対応）
  */
 window.onload = () => {
-  class Confirm {
-    constructor(el) {
-      this.message = el.getAttribute('data-confirm')
-      if (this.message) {
-        el.form.addEventListener('submit', this.confirm.bind(this))
-      } else {
-        console && console.warn('No value specified in `data-confirm`', el)
-      }
-    }
-
-    confirm(e) {
-      if (!window.confirm(this.message)) {
-        e.preventDefault();
-      }
-    }
-  }
-
-  Array.from(document.querySelectorAll('[data-confirm]')).forEach((el) => {
-    new Confirm(el)
-  })
+  document.querySelectorAll('[data-confirm]').forEach(el => new Confirm(el))
 }
