+++ todoapp/app/views/tasks/_form.html.erb
@@ -9,19 +9,19 @@
 
     <%= form_with model: task, local: true, class: 'form-horizontal' do |f| %>
       <div class="form-group row">
-        <%= f.label :title, class: 'col-form-label col-sm-2' %>
+        <%= f.label :title, class: 'col-form-label col-lg-2 col-sm-3' %>
         <%= f.text_field :title, class: 'form-control col-7' %>
       </div>
       <div class="form-group row">
-        <%= f.label :description, class: 'col-form-label col-sm-2' %>
+        <%= f.label :description, class: 'col-form-label col-lg-2 col-sm-3' %>
         <%= f.text_area :description, class: 'form-control col-7' %>
       </div>
       <div class="form-group row">
-        <%= f.label :status, class: 'col-form-label col-sm-2' %>
+        <%= f.label :status, class: 'col-form-label col-lg-2 col-sm-3' %>
         <%= f.select :status, Task.statuses_i18n.invert, {}, { class: 'form-control col-7' } %>
       </div>
       <div class="form-group row">
-        <%= f.label :end_at, class: 'col-form-label col-sm-2' %>
+        <%= f.label :end_at, class: 'col-form-label col-lg-2 col-sm-3' %>
         <%= f.date_field :end_at, min: Time.now.to_date, max: Time.now.to_date + (365 * 100), class: 'form-control col-7' %>
       </div>
